requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.128.0
uvicorn
openai-whisper
numpy
scikit-learn
joblib
torch==2.10.0+cpu
